+++ packages/js/react/src/index.js
@@ -3,7 +3,7 @@
  * Provides hooks and components for React applications.
  */
 
-import LaravelValidator from '../core/LaravelValidator.js';
+import LaravelValidator from '@laravel-client-validation/core';
 
 const defaults = {
   remoteUrl: '/client-validation/validate',